+++ FuryCore/Interfaces/IGameObjectType.cs
@@ -1,0 +1,10 @@
+﻿namespace StardewMods.FuryCore.Interfaces;
+
+using StardewMods.FuryCore.Interfaces.GameObjects;
+
+/// <summary>
+///     Represents any type of <see cref="IGameObject" /> that is accessible to the player.
+/// </summary>
+public interface IGameObjectType
+{
+}